requirements.txt
streamlit==1.28.1
langchain==0.0.350
langchain-community==0.0.10
langchain-text-splitters==0.0.1
sentence-transformers==2.2.2
pinecone-client==2.2.4
transformers==4.35.2
torch==2.1.1
python-dotenv==1.0.0
pypdf==3.17.1
tiktoken==0.5.1 